move caliper-unrelated code from caliper app to ucsd_features
